pyproject.toml
[project]
name = "transformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "kagglehub>=1.0.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pyspellchecker>=0.8.4",
    "pytest>=9.0.2",
    "ruff>=0.15.3",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "ty>=0.0.19",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
	"pytest>=8.2.2",
]
